+++ src/pages/create/create.tsx
@@ -1,14 +1,14 @@
 import React, { useEffect, useRef, useState } from 'react';
 import { useHistory } from 'react-router-dom';
 import { useRecoilState } from 'recoil';
-import { Button, Form, Input, Select, Checkbox, notification } from 'antd';
+import { Button, Checkbox, Form, Input, notification, Select } from 'antd';
 import { FormInstance } from 'antd/lib/form';
 import { walletIdentifierState, walletTempBackupState } from '../../recoil/atom';
 import './create.less';
 import { Wallet } from '../../models/Wallet';
 import { walletService } from '../../service/WalletService';
 import { WalletCreateOptions, WalletCreator } from '../../service/WalletCreator';
-import { DefaultWalletConfigs, NodePorts } from '../../config/StaticConfig';
+import { DefaultWalletConfigs, LedgerWalletMaximum, NodePorts } from '../../config/StaticConfig';
 import logo from '../../assets/logo-products-chain.svg';
 import SuccessModalPopup from '../../components/SuccessModalPopup/SuccessModalPopup';
 import ErrorModalPopup from '../../components/ErrorModalPopup/ErrorModalPopup';
@@ -22,9 +22,11 @@ import SuccessCheckmark from '../../components/SuccessCheckmark/SuccessCheckmark
 import IconLedger from '../../svg/IconLedger';
 import {
   createLedgerDevice,
+  detectLedgerStatus,
   LEDGER_WALLET_TYPE,
   NORMAL_WALLET_TYPE,
 } from '../../service/LedgerService';
+import { TransactionUtils } from '../../utils/TransactionUtils';
 
 const layout = {
   // labelCol: { span: 8 },
@@ -372,6 +374,16 @@ const FormCreate: React.FC<FormCreateProps> = props => {
 
       hwok = true;
     } catch (e) {
+      let message = `Cannot detect any Ledger device`;
+      let description = `Please connect with your Ledger device`;
+      if (walletType === LEDGER_WALLET_TYPE) {
+        const ledgerStatus = await detectLedgerStatus();
+        if (ledgerStatus === 'NORMAL') {
+          message = `No Expert Mode`;
+          description = 'Please ensure that your have enabled Expert mode on your ledger device.';
+        }
+      }
+
       props.setLedgerConnected(false);
 
       await new Promise(resolve => {
@@ -380,10 +392,10 @@ const FormCreate: React.FC<FormCreateProps> = props => {
       props.setIsModalVisible(false);
 
       notification.error({
-        message: `Cannot detect any Ledger device`,
-        description: `Please connect with your Ledger device`,
+        message,
+        description,
         placement: 'topRight',
-        duration: 3,
+        duration: 20,
       });
     }
     await new Promise(resolve => {
@@ -394,6 +406,12 @@ const FormCreate: React.FC<FormCreateProps> = props => {
       onWalletCreateFinishCore();
     }
   };
+
+  const addressIndexValidator = TransactionUtils.rangeValidator(
+    `0`,
+    `${LedgerWalletMaximum}`,
+    `Address index should be between 0 ~ ${LedgerWalletMaximum} (both inclusive)`,
+  );
 
   return (
     <Form
@@ -438,6 +456,7 @@ const FormCreate: React.FC<FormCreateProps> = props => {
             required: true,
             message: 'Please input your address index',
           },
+          addressIndexValidator,
         ]}
         hidden={props.isWalletSelectFieldDisable}
       >