+++ src/service/LedgerService.ts
@@ -3,6 +3,7 @@ import { ISignerProvider } from './signers/SignerProvider';
 import { LedgerWalletSignerProviderNative } from './signers/LedgerWalletSignerProviderNative';
 // eslint-disable-next-line     @typescript-eslint/no-unused-vars
 import { LedgerWalletSignerProviderWebusb } from './signers/LedgerWalletSignerProviderWebusb';
+import { LedgerWalletMaximum } from '../config/StaticConfig';
 
 export const useWebusbForLedger = false;
 export const LEDGER_WALLET_TYPE = 'ledger';
@@ -15,4 +16,20 @@ export function createLedgerDevice(): ISignerProvider {
     signerProvider = new LedgerWalletSignerProviderNative();
   }
   return signerProvider;
+}
+
+// NORMAL
+// EXPERT
+// DISCONNECTED
+export async function detectLedgerStatus(): Promise<String> {
+  try {
+    const device = createLedgerDevice();
+    await device.getPubKey(LedgerWalletMaximum, false);
+    return 'EXPERT';
+  } catch (err) {
+    const findResult = err.toString().search('Conditions not satisfied');
+    if (findResult >= 0) return 'NORMAL';
+
+    return 'DISCONNECTED';
+  }
 }