+++ CHANGELOG.md
@@ -6,6 +6,16 @@ All notable changes to this project will be documented in this file.
 
 *Released*
 
+## [v0.1.6] - 2021-04-14
+
+### Additions
+- Added capability to sort wallets by wallet names and wallet types
+- General configuration capability now supported - Now possible to propagate changes to all wallets of a certain type
+- Ledger, remove 100 wallets index limit - Now up to 2147483647 wallets indexes supported
+  
+### Bug Fixes
+- Fix Ledger disconnection issues
+
 ## [v0.1.5] - 2021-04-08
 
 ### Additions